+++ packages/drizzle/src/queries/parseParams.ts
@@ -9,7 +9,6 @@ import { validOperators } from 'payload/shared'
 import type { DrizzleAdapter, GenericColumn } from '../types.js'
 import type { BuildQueryJoinAliases } from './buildQuery.js'
 
-import { buildAndOrConditions } from './buildAndOrConditions.js'
 import { getTableColumnFromPath } from './getTableColumnFromPath.js'
 import { sanitizeQueryValue } from './sanitizeQueryValue.js'
 
@@ -348,4 +347,47 @@ export function parseParams({
   }
 
   return result
+}
+
+export function buildAndOrConditions({
+  adapter,
+  fields,
+  joins,
+  locale,
+  selectFields,
+  tableName,
+  where,
+}: {
+  adapter: DrizzleAdapter
+  collectionSlug?: string
+  fields: Field[]
+  globalSlug?: string
+  joins: BuildQueryJoinAliases
+  locale?: string
+  selectFields: Record<string, GenericColumn>
+  tableName: string
+  where: Where[]
+}): SQL[] {
+  const completedConditions = []
+  // Loop over all AND / OR operations and add them to the AND / OR query param
+  // Operations should come through as an array
+
+  for (const condition of where) {
+    // If the operation is properly formatted as an object
+    if (typeof condition === 'object') {
+      const result = parseParams({
+        adapter,
+        fields,
+        joins,
+        locale,
+        selectFields,
+        tableName,
+        where: condition,
+      })
+      if (result && Object.keys(result).length > 0) {
+        completedConditions.push(result)
+      }
+    }
+  }
+  return completedConditions
 }
